ml_project/requirements.txt
click==7.1.2
joblib==1.0.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
mypy-extensions==0.4.3
numpy==1.20.2
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0

src~=0.1.0
pytest~=6.2.4
Faker~=8.1.2
setuptools~=56.1.0